requirements.txt
anyio==4.0.0
ariadne==0.17.0
blinker==1.6.3
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.3
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.31.1
typing_extensions==4.8.0
Werkzeug==3.0.0
